requirements.txt
openai
langchain
python-dotenv
fastapi
uvicorn
azure-cli
azure-functions
azure-identity
tiktoken
pymilvus
streamlit
pandas
numpy